requirements-dev.txt
# Development and testing dependencies
pytest>=7,<9
jsonschema>=4,<5
pynacl>=1.5,<2
pyyaml>=6,<7
mavsdk>=2,<3
ruff>=0.4,<1
opencv-python-headless>=4.8,<5
numpy>=1.26,<3
